% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_markers.R
\name{align_markers}
\alias{align_markers}
\title{Title}
\usage{
align_markers(
  markers,
  data_path,
  metadata,
  nchunks,
  ncores,
  iocores,
  vch = 5,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
