% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{align_peaks}
\alias{align_peaks}
\title{Align Peaks in Spectra}
\usage{
align_peaks(
  s,
  tolerance = 0.002,
  ppm = NULL,
  minFreq = NULL,
  reference = NULL,
  labels = NULL,
  return_ref = FALSE,
  return_warp = FALSE,
  ...
)
}
\arguments{
\item{s}{A \code{Spectra} object containing the spectral data to be aligned.}

\item{tolerance}{A numeric value specifying the m/z tolerance for peak matching.
If \code{reference} is \code{NULL}, the same tolerance is used to build an internal reference.}

\item{ppm}{Tolerance deviation in ppm. \code{tolerance} is ignored if \code{ppm} is set.}

\item{minFreq}{A numeric value between 0 and 1 representing the minimum frequency of a peak across spectra for it to be considered a reference peak. Defaults to 0.9 if \code{reference} is \code{NULL}.}

\item{reference}{An optional \link[MALDIquant:MassPeaks-class]{MALDIquant::MassPeaks} object to be used as the reference for alignment. If \code{NULL}, a reference is created based on \code{minFreq}.}

\item{labels}{Optional labels for the reference peaks.}

\item{return_ref}{Logical. If \code{TRUE}, the function returns a list with the aligned spectra and the reference peaks. Defaults to \code{FALSE}.}

\item{return_warp}{Whether warping functions are returned, added as a Spectra varaible}

\item{...}{Additional arguments to control alignment behavior. Supported arguments:
\itemize{
\item \code{allowNoMatches}: Logical, if \code{TRUE}, allows no matches in warping functions.
\item \code{emptyNoMatches}: Logical, if \code{TRUE}, inserts empty values for unmatched peaks in the output.
}}
}
\value{
If \code{return_ref = FALSE}, returns an aligned \code{Spectra} object. If \code{return_ref = TRUE}, returns a list containing the aligned \code{Spectra} object and the reference peaks.
}
\description{
This function aligns peaks across spectra using a specified tolerance using a reference peak list.
See \link[MALDIrppa:alignPeaks]{MALDIrppa::alignPeaks}.
}
\details{
The function first processes the peaks in \code{s}, and if no reference is provided, it generates one based on \code{minFreq}. The function applies warping functions determined by the \code{determineWarpingFunctions} function and then aligns the spectra.
}
\examples{
# Example usage:
# align_peaks(spectra, tolerance = 0.1)
}
