% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_markers.R
\name{align_sample}
\alias{align_sample}
\title{Title}
\usage{
align_sample(s, markers, myby, gauss, laglim, halfWindowSize)
}
\arguments{
\item{s}{MassSpectrum object}

\item{markers}{markers data frame}

\item{laglim}{}
}
\description{
Title
}
