% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_maldiquant.R
\name{asMassPeaksList}
\alias{asMassPeaksList}
\title{Transform list of peak matrices into a list of [MALDIquant::MassPeaks]}
\usage{
asMassPeaksList(pl)
}
\arguments{
\item{pl}{List of peak matrices}
}
\value{
List of [MALDIquant::MassPeaks] objects
}
\description{
Transform list of peak matrices into a list of [MALDIquant::MassPeaks]
}
