% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{atipicality_spectra}
\alias{atipicality_spectra}
\title{Quality control spectra from [MALDIrppa::screenSpectra]
Works on peak matrix rather than on [MALDIquant::MassSpectrum]}
\usage{
atipicality_spectra(x, ...)
}
\arguments{
\item{x}{}
}
\value{
numeric, Atipicality score
}
\description{
Quality control spectra from [MALDIrppa::screenSpectra]
Works on peak matrix rather than on [MALDIquant::MassSpectrum]
}
