% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{atipicalitySpectra}
\alias{atipicalitySpectra}
\title{Identification of potentially low-quality raw mass spectra}
\usage{
atipicalitySpectra(s)
}
\arguments{
\item{object}{[Spectra::Spectra()] object}
}
\value{
A [Spectra::Spectra()] object with \code{QCflag} spectrum variable added
}
\description{
Uses [atipicality_spectra], which implements [MALDIrppa::screenSpectra()] on
peak matrix
}
