% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{baseline_correction}
\alias{baseline_correction}
\title{Baseline subtraction}
\usage{
baseline_correction(
  x,
  int_index = 2,
  keep_bl = TRUE,
  substract_index = 2,
  in_place = FALSE,
  ...
)
}
\arguments{
\item{x}{Peak matrix}

\item{int_index}{Index of the intensity to calculate the baseline}

\item{substract_index}{Index of the intensity column to be substracted.
If NULL, there is no substraction}

\item{in_place}{Replace the intensity in `substract_index` with the baseline
substracted intensity}

\item{...}{Arguments passed to [bl_estim()]}
}
\description{
Baseline subtraction
}
