% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{change_format}
\alias{change_format}
\title{Read MALDI data in a given format at once and export in a different one}
\usage{
change_format(indir, readf, outdir, writef)
}
\arguments{
\item{writef}{}
}
\value{

}
\description{
Read MALDI data in a given format at once and export in a different one
}
