% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{change_format_chunks}
\alias{change_format_chunks}
\title{Change MALDI-TOF data format}
\usage{
change_format_chunks(
  spectra_in,
  spectra_out = NULL,
  outpath = NULL,
  readf = NULL,
  indir = NULL,
  in_ext = NULL,
  writef = NULL,
  sep = "\\t",
  ncores = 4,
  nchunks = 80,
  verbose = FALSE
)
}
\arguments{
\item{spectra_in}{File names or paths to transform (with or without extension)}

\item{spectra_out}{Names of the spectra to go in the output. These can either constitute
filenames and/or spectrumId in a mzML file.}

\item{outpath}{Destination path}

\item{readf}{Input format reading function. One of \code{'fread'}, \code{'table'}
or \code{'mzML'}. \code{'fread'} uses \code{\link[data.table:fread]{data.table::fread()}},
\code{'table'} uses \code{\link[utils:read.table]{utils::read.table()}} and \code{mzML} uses \code{\link[mzR:openMSfile]{mzR::openMSfile()}}.}

\item{indir}{Path to data folder. If \code{NULL}, it assumes \code{spectra_in} contains whole
paths to files.}

\item{in_ext}{Extension of input files}

\item{writef}{Output format writing function. one of 'tab' or 'mzML'.
Determines the destination format of the files.}

\item{sep}{Separator for tsv input files. Default is \code{"\t"}}

\item{ncores}{Number of cores to use}

\item{nchunks}{Number of chunks to split the files into}

\item{verbose}{Print progress bar}
}
\description{
Read MALDI data in tsv (tab) or mzML format in chunks and export in tsv (tab) or mzML.
It is a wrapper interface around reading and writting functions.
}
