% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{check_empty}
\alias{check_empty}
\title{Check is spectra file is empty}
\usage{
check_empty(x, indir)
}
\arguments{
\item{x}{File name}

\item{indir}{Directory to data}
}
\value{
\code{TRUE} if file has any data, \code{FALSE} if it's empty
}
\description{
Check is spectra file is empty
}
