% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{concatenate_spectra}
\alias{concatenate_spectra}
\title{Concatenation of Spectra
Prior concatenation, it applies the processing queue and changes to a
MsBackendDataFrame}
\usage{
concatenate_spectra(x, ...)
}
\arguments{
\item{x}{List of Spectra objects}
}
\description{
Concatenation of Spectra
Prior concatenation, it applies the processing queue and changes to a
MsBackendDataFrame
}
