% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{count_peaks}
\alias{count_peaks}
\title{Count peaks in each spectra}
\usage{
count_peaks(s)
}
\arguments{
\item{s}{\link[Spectra:Spectra]{Spectra::Spectra} object}
}
\value{
A vector of integers with the number of peaks
}
\description{
Count peaks in each spectra
}
