% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{custom_alignPeaks}
\alias{custom_alignPeaks}
\title{Title}
\usage{
custom_alignPeaks(
  l,
  tolerance,
  minFreq = NULL,
  reference = NULL,
  labels = NULL
)
}
\arguments{
\item{th_peaks}{}
}
\description{
Title
}
