% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{export_mzR}
\alias{export_mzR}
\title{Export a matrix into mzML using mzR}
\usage{
export_mzR(x, id, outpath, header_template)
}
\arguments{
\item{x}{Matrix with mz and intensities}

\item{id}{Spectra ID}

\item{template_header}{}
}
\description{
Export a matrix into mzML using mzR
}
