% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{export_tsv}
\alias{export_tsv}
\title{Export list of spectra in multiple \code{"tsv"} files}
\usage{
export_tsv(l, path)
}
\arguments{
\item{l}{List of spectra}

\item{path}{List of paths to write each spectra in \code{l}.}
}
\description{
Export list of spectra in multiple \code{"tsv"} files
}
