% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{extRefPeaks}
\alias{extRefPeaks}
\title{Build reference Spectrum from external sequences}
\usage{
extRefPeaks(
  sequences,
  mc.cores = 4L,
  gpo_only = F,
  ret.object = c("df", "MassPeaks", "Spectra")
)
}
\arguments{
\item{ret.object}{}
}
\description{
Build reference Spectrum from external sequences
}
