% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{filterSNR}
\alias{filterSNR}
\title{Filter peaks in Spectra object by SNR}
\usage{
filterSNR(s, snr)
}
\arguments{
\item{s}{Spectra object}

\item{snr}{signal-to-noise threshold}
}
\value{
Spectra object
}
\description{
Filter peaks in Spectra object by SNR
}
