% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_markers.R
\name{get_max_cor}
\alias{get_max_cor}
\title{Title}
\usage{
get_max_cor(
  markers,
  data_path,
  metadata,
  mat_format,
  nchunks,
  ncores,
  iocores,
  vch = 5,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
