% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{get_spectra_name}
\alias{get_spectra_name}
\title{Extract spectra name from a full path with extensions}
\usage{
get_spectra_name(files)
}
\arguments{
\item{files}{Paths to the individual spectra files including the file name and extension}
}
\value{
A vector with sample names
}
\description{
Spectra name has the format "samplename_replicate"
}
