% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getPeptPeaks}
\alias{getPeptPeaks}
\title{Title}
\usage{
getPeptPeaks(s, ...)
}
\arguments{
\item{s}{\code{\link[MALDIquant]{MassSpectrum}} object}

\item{...}{Parameters passed to the different preprocessing functions.
See \code{\link{preprocess_data}}}
}
\value{
peaks list, with mass, intensity and s2n
}
\description{
Title
}
