% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{import_tsv}
\alias{import_tsv}
\title{Import data in tsv format}
\usage{
import_tsv(f, sep = "\\t")
}
\arguments{
\item{f}{Path to file}

\item{sep}{Separator. Default is \code{"\t"}}
}
\value{
Matrix with mz and intensities
}
\description{
Uses \link[data.table:fread]{data.table::fread}
}
