% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{initialize_mzml_header}
\alias{initialize_mzml_header}
\title{Returns mzML header data for [mzR]}
\usage{
initialize_mzml_header()
}
\value{
A one row data.frame
}
\description{
Returns mzML header data for [mzR]
}
