% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{int_ref_peaks}
\alias{int_ref_peaks}
\title{Build internal reference spectrum}
\usage{
int_ref_peaks(
  mpl,
  method = c("strict", "relaxed"),
  minFreq = 0.9,
  tolerance = 0.002,
  labels = NULL,
  ret.object = c("Spectra", "MassPeaks", "df")
)
}
\arguments{
\item{mpl}{MassPeaksList object}

\item{labels}{}
}
\description{
Build internal reference spectrum
}
