% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{intensity_matrix}
\alias{intensity_matrix}
\title{Intensity matrix}
\usage{
intensity_matrix(s, specvar_to_names = NULL)
}
\arguments{
\item{s}{Spectra or MassPeaks object}

\item{specvar_to_names}{Spectra variable to use as rownames of the
intensity matrix}
}
\value{
\code{matrix} with samples as rows and masses as columns, filled with
intensity values
}
\description{
Turns the list of peaks into an intensity matrix with samples/spectra as rows
and mz values as columns.
}
