% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{intRefPeaks}
\alias{intRefPeaks}
\title{Title}
\usage{
intRefPeaks(
  l,
  method = c("strict", "relaxed"),
  minFreq = 0.9,
  tolerance = 0.002,
  labels = NULL
)
}
\arguments{
\item{labels}{}
}
\description{
Title
}
