% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeReplicates}
\alias{mergeReplicates}
\title{Title}
\usage{
mergeReplicates(
  l,
  labels,
  names,
  method = c("mean", "median", "sum"),
  ignore.na = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
