% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{modelLocalBG}
\alias{modelLocalBG}
\title{modelLocalBG}
\usage{
modelLocalBG(p, full_m, mass_range, bg_cutoff)
}
\arguments{
\item{p}{Matrix, mass-intensity pair}

\item{mass_range}{Mass window to both sides of \code{p} to be considered for backgroun modelling}

\item{bg_cutoff}{The peaks within the mass range with intensity below the \code{bg_cutoff} quantile
are considered for background modelling. \code{bg_cutoff=1} keeps all peaks
and \code{bg_cutoff=0.5} would only keep the bottom half.}

\item{m}{Full matrix of mass-intensity pairs}

\item{l_cutoff}{Likelihood threshold or p-value. Peaks with a probability of being modelled as
background noise higher than this are filtered out.}
}
\value{

}
\description{
modelLocalBG
}
