% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{monoisotopic_peaks}
\alias{monoisotopic_peaks}
\title{Find monoisotopic peaks
It is a wrapper on [MALDIquant::monoisotopicPeaks] to work with Spectra objects}
\usage{
monoisotopic_peaks(s, ppm, ...)
}
\arguments{
\item{s}{Spectra object}

\item{ppm}{tolerance in parts per million}

\item{...}{\itemize{
   \item minCor
   \item distance
   \item size
}}
}
\description{
Find monoisotopic peaks
It is a wrapper on [MALDIquant::monoisotopicPeaks] to work with Spectra objects
}
