% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normalize_spectra}
\alias{normalize_spectra}
\title{Normalize peaks matrix}
\usage{
normalize_spectra(x, scaleFactor = 1, func = NULL, ...)
}
\arguments{
\item{x}{Peaks matrix}

\item{scaleFactor}{Scaling factor}

\item{func}{Function to be applied to the intensities}

\item{...}{Currently not in use, required by \link[Spectra:Spectra]{Spectra::addProcessing,Spectra-method}.}
}
\value{
Peaks matrix
}
\description{
Normalize peaks matrix
}
