% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normalizeSpectra}
\alias{normalizeSpectra}
\title{Normalize spectra object}
\usage{
normalizeSpectra(s, scale_f)
}
\arguments{
\item{s}{\link[Spectra:Spectra]{Spectra::Spectra} object}

\item{scale_f}{spectra variable to use as normalization factor or a function
to be applied to spectra intensities}
}
\value{
\link[Spectra:Spectra]{Spectra::Spectra} object
}
\description{
Normalize spectra object
}
