% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{peakDetection}
\alias{peakDetection}
\title{Peak detection on Spectra object}
\usage{
peakDetection(s, apply_queue = FALSE, ...)
}
\arguments{
\item{s}{Spectra object}

\item{apply_queue}{logical, whether to apply the processing queue}

\item{...}{Arguments passed to [peak_detection] and [apply_preprocess]}
}
\value{
A Spectra object
}
\description{
Peak detection on Spectra object
}
