% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_markers.R
\name{pept_fly}
\alias{pept_fly}
\title{Title}
\usage{
pept_fly(
  markers,
  peaks,
  peaksby = NULL,
  tolerance = 0.002,
  match_tol = 0.6,
  aug_deam = F
)
}
\arguments{
\item{aug_deam}{}
}
\description{
Title
}
