% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{peptidePseudoClusters}
\alias{peptidePseudoClusters}
\title{peptidePseudoClusters}
\usage{
peptidePseudoClusters(x, masses, tol, n_isopeaks, min_isopeaks)
}
\arguments{
\item{x}{\code{\link[MALDIquant]{MassPeaks}} object}

\item{masses}{1D Vector of isotopic masses to search.
See examples for how to generate from a vector of single peptide masses}

\item{min_isopeaks}{}
}
\value{
data.frame with mass, intensity and s2n values for pseudo-isotopic
        clusters
}
\description{
Extracts pseudo envelopes from given peptide masses
}
\examples{
masses = c(100, 200, 300)
n_isopeaks = 5
masses = matrix(masses, nrow = n_isopeaks, ncol = length(masses), byrow = TRUE)
d = 1.00235
masses = masses + (d * 0L:(n_isopeaks - 1L))
masses = sort(masses)
print(masses)
}
