% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{plot_platemap}
\alias{plot_platemap}
\title{Plot spectra value in platemap}
\usage{
plot_platemap(
  pl,
  plt_var,
  max_row = NULL,
  max_col = NULL,
  plate_id = "",
  n_viridis = 100L
)
}
\arguments{
\item{pl}{platemap with at least two columns, 'spot' and \code{plt_var}, a numeric value to plot}

\item{plt_var}{Nanme of the variable from \code{pl} to plot}

\item{max_row}{Maximum number of rows in the platemap}

\item{max_col}{Maximum number of columns in the platemap}
}
\description{
Plot spectra value in platemap
}
