% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{plot_preprocess}
\alias{plot_preprocess}
\title{Plot Spectra in different preprocessing stages}
\usage{
plot_preprocess(
  l,
  p = NULL,
  scale_factor = 1,
  mzrange = c(800, 4000),
  prep_labels = c("Raw", "Smoothed", "Baseline corrected", "Peaks"),
  n_col = 2
)
}
\arguments{
\item{l}{List of Spectra objects. Each Spectra is at a different state in the
preprocessing.}

\item{p}{Spectra object with peaks}

\item{scale_factor}{Vector of scale factors used to normalize peak with respect
to the rest of the spectra.}

\item{mzrange}{mz range to limit plot}
}
\description{
Plot Spectra in different preprocessing stages
}
