% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepFun}
\alias{prepFun}
\title{prepf
Pre-load spectra preprocessing function with arguments using a closure}
\usage{
prepFun(FUN, ...)
}
\arguments{
\item{FUN}{Function to be loaded with argument in \code{...}}

\item{...}{Arguments passed to \code{FUN}}
}
\value{
A closure with the preprocessing function \code{FUN} and preloaded arguments
}
\description{
prepf
Pre-load spectra preprocessing function with arguments using a closure
}
