% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_data}
\alias{preprocess_data}
\title{preprocess_data}
\usage{
preprocess_data(
  indir,
  readf = c("fread", "table", "mzml"),
  chunks = 50,
  prepf = NULL,
  ncores = NULL,
  iocores = 1
)
}
\arguments{
\item{indir}{Folder containing spectra.}

\item{readf}{A string value. choose function to use to read spectra.
Currently restricted to one of "fread", "table" or "mzml"}

\item{chunks}{Number of chunks all the spectra should be divided into for reading and
processing. If all spectra is loaded and processed at once, i.e. chunks=1,
it can overload RAM memory. If chunks>1 data is loaded and processed to the
much lighter list of peaks in chunks batches.}

\item{prepf}{Custom preprocessing function}

\item{ncores}{Number of cores used for the preprocessing of spectra. The cores will work in
parallel with the different spectra within a chunk.}

\item{iocores}{Number of cores used for I/O operations. For some systems I/O operations involving
multiple cores reading or writing at the same time from disk increases time.}
}
\value{
A list of objects returned by prepf
}
\description{
Performs smoothening, baseline removal and peak detection on MALDI samples.
}
