% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{print_progress}
\alias{print_progress}
\title{Print message only if verbose}
\usage{
print_progress(msg, verbose)
}
\arguments{
\item{msg}{Message to print}

\item{verbose}{logical, whether to print the message or not}
}
\description{
Print message only if verbose
}
\examples{
verbose = TRUE
print_progress('', verbose)
}
