% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{rw_chunk_tsv}
\alias{rw_chunk_tsv}
\title{Reads data in writes in tsv format}
\usage{
rw_chunk_tsv(x, ch, indir, read_f, outpath, in_ext, pb)
}
\description{
Internal use only. Use \link{change_format_chunks} instead
}
