% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{separate_sample_replicate}
\alias{separate_sample_replicate}
\title{Separate file names into sample and replicate}
\usage{
separate_sample_replicate(spectra_names, sep = "_")
}
\arguments{
\item{spectra_names}{Spectra names}

\item{sep}{Separator of sample name and replicate. Default is '_'}
}
\description{
Separate file names into sample and replicate
}
