% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{separate_sample_replicate_df}
\alias{separate_sample_replicate_df}
\title{Separate spectra_name column into sample and replicate}
\usage{
separate_sample_replicate_df(sp, sample_sep = "_", repl_sep = "_")
}
\arguments{
\item{sp}{data.frame with data, in which the spectra name is in 'spectra_name'
column}

\item{repl_seq}{}
}
\description{
Separate spectra_name column into sample and replicate
}
