% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{smoothSpectra}
\alias{smoothSpectra}
\title{Smooth spectra}
\usage{
smoothSpectra(s, ...)
}
\arguments{
\item{s}{Spectra object}

\item{...}{}
}
\value{
A Spectra object
}
\description{
Smooth spectra
}
