% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_markers.R
\name{ts_align}
\alias{ts_align}
\title{Title}
\usage{
ts_align(
  ts,
  data,
  txlim,
  laglim,
  myby,
  gauss = NA,
  normlim = NA,
  ccylim = c(-0.1, 0.5)
)
}
\arguments{
\item{ccylim}{}
}
\description{
Title
}
