% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{align_peaks}
\alias{align_peaks}
\title{Align peaks}
\usage{
align_peaks(s, tolerance, minFreq = NULL, reference = NULL, labels = NULL, ...)
}
\arguments{
\item{s}{Spectra object. Must be centroided}

\item{...}{}
}
\value{
A Spectra object with MsBackendDataFrame
}
\description{
Prior aligning, all processing queue is applied
}
