% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{apply_preprocess}
\alias{apply_preprocess}
\title{Applies processing queue to Spectra}
\usage{
apply_preprocess(s, write_data = FALSE, file = NULL)
}
\arguments{
\item{s}{Spectra object}

\item{write_data}{logical. Whether to write spectra data into a file
This will apply processing, export spectra into mzML file(s) and
reload into a new Spectra object with MsBackendMzR. This is useful
if we want to apply preprocessing and the resulting object is still
too big to keep in memory, either because there are too many spectra
or because we still haven't picked peaks.}

\item{file}{If `write_data` is `TRUE`, file or files to write spectra.
IF there is one file, all spectra is saved on the same file.
If multiple files, there needs to be one per spectra.}
}
\value{
A Spectra object with either [Spectra::MsBackendMzR] or
[Spectra::MsBackendDataFrame] backends
}
\description{
Applies processing queue to Spectra
}
