% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{baselineCorrection}
\alias{baselineCorrection}
\title{Apply baseline correction to Spectra object}
\usage{
baselineCorrection(s, ...)
}
\arguments{
\item{s}{Spectra object}

\item{...}{}
}
\value{
Spectra object
}
\description{
Apply baseline correction to Spectra object
}
