% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{bin_peaks}
\alias{bin_peaks}
\title{Bin peaks across spectra}
\usage{
bin_peaks(s, method = c("strict", "relaxed", "reference"), tolerance = 0.002)
}
\arguments{
\item{tolerance}{}
}
\description{
Bin peaks across spectra
}
