% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{change_format_chunks}
\alias{change_format_chunks}
\title{Read MALDI data in a given format in chunks and export in a different one}
\usage{
change_format_chunks(
  spectra_names,
  indir,
  in_fmt,
  readf,
  outpath,
  writef,
  sep = "\\t",
  mc.cores = 4,
  nchunks = 80
)
}
\arguments{
\item{in_fmt}{}

\item{readf}{One of 'fread', 'table' or 'mzML'}

\item{writef}{One of 'tab' or 'mzML'}
}
\description{
Read MALDI data in a given format in chunks and export in a different one
}
