% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{collect_triplicates}
\alias{collect_triplicates}
\title{Arrange samples from different platemaps into a common folder}
\usage{
collect_triplicates(
  platemap,
  basepath = "",
  ext = NULL,
  outfolder = NULL,
  keep_dupl = T,
  keep_incomplete_tripl = T,
  dry_run = F
)
}
\arguments{
\item{platemap}{Data frame mapping plate and coordinate to sample name and replicate.
It must contain the following columns:
\itemize{
  \item folder: main folder of group of plates.
  \item subfolder: subfolder from a specific plate where spetra is stored.
  \item basename: prefix of samples in the folder that is preceeded by coordinates.
  \item sample_name: (preferably) unique sample name.
  \item spot1, spot2, spot3: coordinates of each of the replicates.
  \item date: the date each plate was spotted or analyzed.
  There can be more than 3 replicates, but it always need to have this name: "spot#"
}}

\item{basepath}{Path to plates}

\item{ext}{File extension. E.g. mzML, csv, txt, tab}

\item{outfolder}{Destination path for renamed spectra}

\item{keep_dupl}{Logical. Should duplicate samples be kept? If so, an suffix is added.
The suffix has this form: "__#".}

\item{keep_incomplete_tripl}{Logical. When samples come in triplicates, whether incomplete triplicates
should be removed}

\item{dry_run}{Don't copy spectra, just return an updated platemap, with possible duplicates
included, gathered and with input and output filenames}
}
\value{
A data frame with the modified platemap. Each replicate is now 1 row. Duplicated
samples have the "__#" suffix. Duplicated samples are flagged: \code{dupl} will flag
all duplicates but the last one analyzed by \code{date} in the platemap. \code{dupl_all}
will flag all duplicated samples.
Path for each sample to the original plate location is included in a column.
}
\description{
ZooMS MALDI spectra from the same plate are usually stored in a separate
folder with the plate coordinate as file name.
Using a table that maps plateID and coordinate to sample name and replicate,
this function copies all the spectra into the same folder with appropriate
sample name and replicate number.
}
\examples{
\dontrun{
platemap = read_csv('path/to/platemap.csv')
platemap_1 = collect_triplicates(
  platemap,
  basepath = 'path/to/platefolder', ext = "txt",
  outfolder = 'path/to/samples',
  keep_dupl = T, keep_incomplete_tripl = T, dry_run=F)
}
}
