% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{collect_triplicates}
\alias{collect_triplicates}
\title{Title}
\usage{
collect_triplicates(
  file,
  basepath,
  ext,
  readf,
  outpath,
  writef,
  keep_dupl,
  keep_incomplete_tripl
)
}
\arguments{
\item{basepath}{}
}
\value{

}
\description{
Title
}
