% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{collect_triplicates}
\alias{collect_triplicates}
\title{Title}
\usage{
collect_triplicates(
  platemap,
  basepath,
  ext,
  outfolder,
  keep_dupl,
  keep_incomplete_tripl,
  dry_run = F
)
}
\arguments{
\item{dry_run}{Don't copy spectra, just return an updated platemap, with possible duplicates
included, gathered and with input and output filenames}
}
\value{

}
\description{
Title
}
