% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{collect_triplicates}
\alias{collect_triplicates}
\title{Arrange samples from different platemaps into a common folder}
\usage{
collect_triplicates(
  platemap,
  in_pivot = c("wide", "long"),
  basepath = "",
  ext = NULL,
  outfolder = NULL,
  out_pivot = c("long", "wide"),
  keep_dupl = T,
  keep_incomplete_tripl = T,
  ext_in_folder = FALSE,
  dry_run = F
)
}
\arguments{
\item{platemap}{Data frame mapping plate and coordinate to sample name and replicate.
It can contain the following columns:
\itemize{
  \item folder: main folder of group of plates.
  \item subfolder: subfolder from a specific plate where spetra is stored.
  \item basename: prefix of samples in the folder that is preceeded by coordinates.
  \item sample_name: (preferably) unique sample name.
  \item spot1, spot2, spot3: (for wide format) coordinates of each of the replicates.
        There can be more than 3 replicates, but the column name always need
        to have this format: "spot#"
  \item spot: (for long format), coordinates of the replicate.
  \item date (optionl): the date each plate was spotted or analyzed.
}}

\item{in_pivot}{Whether the platemap is in "wide" or "long" format.
In the wide format, each row is one sample and the coordinates of the
replicates are in different columns as spot1, spot2, ...
In the long format, each row is an individual replicate. In this case, the
platemap must contain a column called "replicate" with the replicate number
and the column with the coordinates should be called "spot".
Long format assumes no duplicates.}

\item{basepath}{Path to plates}

\item{ext}{File extension. E.g. mzML, csv, txt, tab}

\item{outfolder}{Destination path for renamed spectra.}

\item{out_pivot}{Whether to return the platemap in "wide" or "long" format.}

\item{keep_dupl}{Logical. Should duplicate samples/replicates be kept? If so, a suffix is added.
The suffix has this form: "__#".}

\item{keep_incomplete_tripl}{Logical. When samples come in triplicates, whether incomplete triplicates
should be skipped.}

\item{ext_in_folder}{Logical. Whether the platemap folder name has the extension in it, e.g.
20180727_SS08.txt where SS08 is the plate name.}

\item{dry_run}{Logical. Don't copy spectra, just return an updated platemap, with possible duplicates
included, in long format and with input and output filenames}
}
\value{
A data.frame with the modified platemap in long format, so each replicate is
now in 1 row. Duplicated samples, if kept, have the "__#" suffix.
Duplicated samples are flagged with \code{dupl}. All duplicates but the last
one, by \code{date}, in the platemap are flagged.
Column \code{dupl_all} will flag all duplicated samples.
Path for each sample to the original plate location is included in a column.
}
\description{
ZooMS MALDI spectra from the same plate are usually stored in a separate
folder with the plate coordinate as file name.
Using a table that maps plate ID and coordinates to sample name and replicate,
this function copies all the spectra into the same folder with appropriate
sample name and replicate number.
}
\examples{
\dontrun{
platemap = read_csv('path/to/platemap.csv')
platemap_1 = collect_triplicates(
  platemap,
  basepath = 'path/to/platefolder', ext = "txt",
  outfolder = 'path/to/samples',
  keep_dupl = T, keep_incomplete_tripl = T, dry_run=F)
}
}
