% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{concatenate_spectra}
\alias{concatenate_spectra}
\title{Concatenation of Spectra}
\usage{
concatenate_spectra(x, ...)
}
\arguments{
\item{x}{List of \link[Spectra:Spectra]{Spectra::Spectra} object}

\item{...}{Parameters for \link{apply_preprocess}}
}
\value{
A single \link[Spectra:Spectra]{Spectra::Spectra} object
}
\description{
Prior concatenation, it applies the processing queue and changes to a
MsBackendDataFrame
}
