% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{export_mzml}
\alias{export_mzml}
\title{Export a matrix into mzML using [mzR::writeMSdata]}
\usage{
export_mzml(x, outpath, header)
}
\arguments{
\item{x}{Matrix with mz and intensities}

\item{outpath}{Path to output folder}

\item{id}{Spectra ID}

\item{idx}{Scan number of index}

\item{header_template}{Header template to complete with spectra data and id}
}
\description{
Export a matrix into mzML using [mzR::writeMSdata]
}
