% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{ext_ref_peaks}
\alias{ext_ref_peaks}
\title{Build reference Spectrum from external sequences}
\usage{
ext_ref_peaks(
  sequences,
  mc.cores = 4L,
  gpo_only = F,
  ret.object = c("df", "MassPeaks", "Spectra"),
  non_deam = FALSE
)
}
\arguments{
\item{non_deam}{}
}
\description{
Build reference Spectrum from external sequences
}
