% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{filter_peaks}
\alias{filter_peaks}
\title{Filter peaks}
\usage{
filter_peaks(s, minFrequency, minNumber, labels, mergeWhitelists = FALSE)
}
\arguments{
\item{mergeWhitelists}{}
}
\description{
Filter peaks
}
