% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{filter_snr}
\alias{filter_snr}
\title{Filter by SNR threshold}
\usage{
filter_snr(x, snr, ...)
}
\arguments{
\item{x}{peaks matrix}

\item{snr}{signal-to-noise threshold}

\item{...}{Currently not in use, required by \link[Spectra:Spectra]{Spectra::addProcessing,Spectra-method}.}
}
\value{
peaks matrix
}
\description{
Filter by SNR threshold
}
