% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{get_spectra_name}
\alias{get_spectra_name}
\title{Extract spectra name from dataOrigin slot of Spectra object}
\usage{
get_spectra_name(files)
}
\arguments{
\item{files}{Path to a individual spectrum file including the file name and extension}
}
\description{
Extract spectra name from dataOrigin slot of Spectra object
}
