% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{import_table}
\alias{import_table}
\title{Import data in tsv format}
\usage{
import_table(f, sep = "")
}
\arguments{
\item{f}{Path to file}

\item{sep}{Separator. Default is \code{''}, which for [utils::read.table()] can
be one or more white spaces or tabs. If you're certain the separator
is exactly \code{"\t"}, [import_tsv()] is a faster option.}
}
\value{
A data.frame with mz and intensity values
}
\description{
Uses [utils::read.table]
}
