% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{intensity_matrix}
\alias{intensity_matrix}
\title{Intensity matrix}
\usage{
intensity_matrix(s, specvar_to_names)
}
\arguments{
\item{s}{Spectra object}

\item{specvar_to_names}{}
}
\description{
Turns the list of peaks into an intensity matrix with samples/spectra as rows
and mz values as columns.
}
