% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{intRefPeaks}
\alias{intRefPeaks}
\title{Build internal reference spectrum}
\usage{
intRefPeaks(
  mpl,
  method = c("strict", "relaxed"),
  minFreq = 0.9,
  tolerance = 0.002,
  labels = NULL
)
}
\arguments{
\item{mpl}{MassPeaksList object}

\item{labels}{}
}
\description{
Build internal reference spectrum
}
