% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{model_local_bg}
\alias{model_local_bg}
\title{Model local backgroung noise around peak}
\usage{
model_local_bg(p, bg)
}
\arguments{
\item{p}{Peak as one-row matrix, mass-intensity pair}

\item{m}{Full matrix of mass-intensity pairs}

\item{mass_range}{Mass window to both sides of \code{p} to be considered for backgroun modelling}

\item{bg_cutoff}{The peaks within the mass range with intensity below the \code{bg_cutoff} quantile
are considered for background modelling. \code{bg_cutoff=1} keeps all peaks
and \code{bg_cutoff=0.5} would only keep the bottom half.}

\item{l_cutoff}{Likelihood threshold or p-value. Peaks with a probability of being modelled as
background noise higher than this are filtered out.}
}
\value{
One-row matrix with p-value, ie., probability of a peak with a higher intensity,
normal distribution fitting log-likelihood and p-value.
}
\description{
Model local backgroung noise around peak
}
