% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{model_local_bg}
\alias{model_local_bg}
\title{Title}
\usage{
model_local_bg(p, full_m, mass_range, bg_cutoff)
}
\arguments{
\item{bg_cutoff}{bottom fraction to keep}

\item{l_cutoff}{}
}
\value{

}
\description{
Title
}
