% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{monoisotopic_peaks}
\alias{monoisotopic_peaks}
\title{Find monoisotopic peaks
It is a wrapper on \link[MALDIquant:monoisotopicPeaks-methods]{MALDIquant::monoisotopicPeaks} to work with Spectra objects}
\usage{
monoisotopic_peaks(s, tolerance, ppm, ...)
}
\arguments{
\item{s}{Spectra object}

\item{tolerance}{tolerance between observed peaks and the expected pseudo cluster
using \code{distance} between peaks.}

\item{ppm}{tolerance in parts per million. If use, \code{tolerance} is ignored.}

\item{...}{Arguments passed to \link[MALDIquant:monoisotopicPeaks-methods]{MALDIquant::monoisotopicPeaks}
\itemize{
\item 'minCor': \code{double}, minimum correlation between the experimental pseudo isotopic pattern
and the theoretical one generated by the Poisson model.
\item 'distance': mass distance between the peaks in the pattern. Default is 1.00235
\item 'size': sizes of isotopic patterns to look for. Using the default,
\code{3L:10L}, it will look for patterns of size 10 prefferably and decrease it down
to 3.
}}
}
\value{
\link[Spectra:Spectra]{Spectra::Spectra} object with picked monoisotopic peaks
}
\description{
Find monoisotopic peaks
It is a wrapper on \link[MALDIquant:monoisotopicPeaks-methods]{MALDIquant::monoisotopicPeaks} to work with Spectra objects
}
