% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peaks.R
\name{monoisotopic_peaks}
\alias{monoisotopic_peaks}
\title{Find monoisotopic peaks
It is a wrapper on [MALDIquant::monoisotopicPeaks] to work with Spectra objects}
\usage{
monoisotopic_peaks(s, ...)
}
\arguments{
\item{size}{}
}
\description{
Find monoisotopic peaks
It is a wrapper on [MALDIquant::monoisotopicPeaks] to work with Spectra objects
}
