% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normalize_spectra}
\alias{normalize_spectra}
\title{Normalize peak matrix}
\usage{
normalize_spectra(x, scaleFactor = 1, func = NULL, ...)
}
\arguments{
\item{x}{Peak matrix}

\item{scaleFactor}{Scaling factor}

\item{func}{Function to be applied to the intensities}
}
\description{
Normalize peak matrix
}
