% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{peak_detection}
\alias{peak_detection}
\title{Title}
\usage{
peak_detection(
  x,
  halfWindowSize = 2L,
  method = c("MAD", "SuperSmoother"),
  snr = 0L,
  k = 0L,
  descending = FALSE,
  threshold = 0,
  int_index = 2,
  ...
)
}
\arguments{
\item{int_index}{}
}
\description{
Title
}
