% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{peak_detection}
\alias{peak_detection}
\title{Peak detection}
\usage{
peak_detection(
  x,
  halfWindowSize = 2L,
  method = c("MAD", "SuperSmoother"),
  snr = 0L,
  k = 0L,
  descending = FALSE,
  threshold = 0,
  int_index = 2,
  ...
)
}
\arguments{
\item{x}{Peak matrix}

\item{halfWindowSize}{Half size of the sliding window used to detect local
maxima.}

\item{method}{Method for noise estimation. Either \code{"SuperSmoother"} (default)
or \code{"MAD"}}

\item{snr}{Signal-to-noise ratio threshold above which a peak is picked}

\item{k}{Parameter for [MsCoreUtils::refineCentroids]. Values to the left and right
of a peak to be considered to refine the peak mz position}

\item{descending}{Parameter for [MsCoreUtils::refineCentroids]. Only values between
nearest valleys are used.}

\item{threshold}{Parameter for [MsCoreUtils::refineCentroids], a proportion.
Only values above this fraction of the maxima are used.}

\item{int_index}{Index of the intensity in the peak matrix to be used for peak
detection.}

\item{...}{Currently not in use, required by [Spectra::addProcessing].}
}
\value{
Peak matrix, only containing peaks that fullfill all criteria.
}
\description{
It implements [Spectra::pickPeaks()], but allows using intensities in a given
column in the peak matrix
}
