% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{peakLocalNoiseModel}
\alias{peakLocalNoiseModel}
\title{Filter a peak list based on likelihood that peaks are above background local
noise}
\usage{
peakLocalNoiseModel(s, apply_queue = FALSE, calc_frac = FALSE, ...)
}
\arguments{
\item{s}{Spectra object}

\item{...}{arguments to [peaks_local_bg]}

\item{calc_frac_kept}{logical. Whether to calculate fraction of peaks kept
after peak local noise model filtering}
}
\description{
Filter a peak list based on likelihood that peaks are above background local
noise
}
