% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{peptide_pseudo_clusters}
\alias{peptide_pseudo_clusters}
\title{peptidePseudoClusters}
\usage{
peptide_pseudo_clusters(x, mono_masses, tol, n_isopeaks, min_isopeaks, ...)
}
\arguments{
\item{x}{Peaks matrix object}

\item{mono_masses}{Monoisotopic masses}

\item{...}{}
}
\value{
Peaks matrix with isotopic masses and NA as place holders for when
there is no match
}
\description{
Extracts pseudo envelopes from given monoisotopic peptide masses
It just checks whether the monoisotopic peaks and subsequent peaks
at distance of 1.00235 are present, between min_isopeaks and n_isopeaks.
It also check there are no gaps. It doesn't check the isotopic envelop shape.
}
\examples{
masses = c(100, 200, 300)
n_isopeaks = 5
masses = matrix(masses, nrow = n_isopeaks, ncol = length(masses), byrow = TRUE)
d = 1.00235
masses = masses + (d * 0L:(n_isopeaks - 1L))
masses = sort(masses)
print(masses)
}
