% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{peptidePseudoClusters}
\alias{peptidePseudoClusters}
\title{peptidePseudoClusters}
\usage{
peptidePseudoClusters(x, ...)
}
\arguments{
\item{x}{MassPeaks object}

\item{...}{}
}
\value{
data.frame with mass, intensity and s2n values for pseudo-isotopic
        clusters
}
\description{
Extracts pseudo envelopes from given peptide masses
}
