% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemap.R
\name{plot_platemap}
\alias{plot_platemap}
\title{Plot spectra value in platemap}
\usage{
plot_platemap(
  pl,
  plt_var,
  max_row = NULL,
  max_col = NULL,
  plate_id = "",
  n_viridis = 100L
)
}
\arguments{
\item{pl}{platemap with two columns, spot and a numeric value to plot}

\item{max_col}{}
}
\description{
Plot spectra value in platemap
}
