% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessData}
\alias{preprocessData}
\title{preprocessData}
\usage{
preprocessData(
  indir,
  readf = c("fread", "table", "mzml"),
  sep = NULL,
  nchunks = 50,
  prepf = NULL,
  spectra = NULL,
  ncores = NULL,
  iocores = 1,
  vch = 5
)
}
\arguments{
\item{indir}{Folder containing spectra.}

\item{readf}{A string value. choose function to use to read spectra.
Currently restricted to one of "fread", "table" or "mzml"}

\item{sep}{Separator character for input files}

\item{nchunks}{Number of chunks all the spectra should be divided into for reading and
processing. If all spectra is loaded and processed at once, i.e. nchunks=1,
it can overload RAM memory. If nchunks>1 data is loaded and processed to the
much lighter list of peaks in chunks batches.}

\item{prepf}{Custom preprocessing function}

\item{spectra}{Subset of spectra from \code{indir} to analyze. Without extension.
If \code{NULL} (default), all spectra is processed.}

\item{ncores}{Number of cores used for the preprocessing of spectra. The cores will work in
parallel with the different spectra within a chunk.}

\item{iocores}{Number of cores used for I/O operations. For some systems I/O operations involving
multiple cores reading or writing at the same time from disk increases time.}

\item{vch}{Every how many chuncks progress is reported}
}
\value{
A list of objects returned by prepf
}
\description{
Performs smoothening, baseline removal and peak detection on MALDI samples.
}
\details{
When using \code{"fread"} as the reading function, we can specify the separator \code{sep}.
\code{"table"} is only advised for cases where the separator is variable and
can be one or more spaces or tabs. See \code{\link[utils]{read.table}}.
\code{"fread"} uses internally \code{\link[data.table]{fread}}, which is faster.
\code{"mzml"} uses internally \code{\link[MALDIquantForeign]{importMzMl}}
}
