% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spectra.R
\name{rw_chunk_mzml}
\alias{rw_chunk_mzml}
\title{Reads data in writes in mzML format}
\usage{
rw_chunk_mzml(
  x,
  ch,
  in_files,
  out_files,
  out_names,
  read_f,
  many_to_one,
  add_ext,
  BPPARAM
)
}
\description{
Internal use only. Use \link{change_format_chunks} instead
}
