% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{smooth}
\alias{smooth}
\title{Title}
\usage{
smooth(
  x,
  method = c("MovingAverage", "WeightedMovingAverage", "SavitzkyGolay"),
  hws = 4L,
  k = 3L,
  int_index = 2,
  in_place = FALSE,
  ...
)
}
\arguments{
\item{k}{}
}
\description{
Title
}
