% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{smoothSpectra}
\alias{smoothSpectra}
\title{Smooth spectra}
\usage{
smoothSpectra(s, ...)
}
\arguments{
\item{s}{\link[Spectra:Spectra]{Spectra::Spectra} object}

\item{...}{Parameters passed to \link{smooth}}
}
\value{
A \link[Spectra:Spectra]{Spectra::Spectra} object
}
\description{
Smooth spectra
}
